requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.4
django-auth-ldap==5.1.0
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
ldap3==2.9.1
MarkupSafe==3.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-ldap==3.4.4
sqlparse==0.5.3
typing_extensions==4.12.2
Werkzeug==3.1.3
python-decouple==3.8
